# docs/swagger.yaml
definitions:
  models.RetStruct:
    properties:
      cunt:
        format: int64
        type: integer
      name:
        type: string
    type: object
  models.Subscription:
    properties:
      end_date:
        description: '“start_date”: “07-2025”'
        type: string
      price:
        description: '“price”: 400,'
        type: integer
      service_name:
        description: “Yandex Plus”,
        type: string
      start_date:
        description: '“start_date”: “07-2025”'
        type: string
      user_id:
        description: '“user_id”: “60601fee-2bf1-4721-ae6f-7636e79a0cba”,'
        type: string
    type: object
info:
  contact: {}
  description: API для управления подписками
  title: Subscription Service API
  version: "1.0"
paths:
  /:
    get:
      description: Checks if database connection is alive
      produces:
      - application/json
      responses:
        "200":
          description: Database is reachable
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Database connection error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Database health check
  /add:
    post:
      consumes:
      - application/json
      description: Add a new subscription to the database
      parameters:
      - description: Subscription data
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/models.Subscription'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RetStruct'
        "400":
          description: Validation error
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      summary: Create a new subscription
  /delete:
    delete:
      consumes:
      - application/json
      description: Удаляет подписку по переданным данным
      parameters:
      - description: Данные подписки для удаления
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/models.Subscription'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RetStruct'
        "400":
          description: Неверный формат запроса
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Удаление подписки
  /list:
    get:
      description: Возвращает полный список всех подписок из базы данных
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Subscription'
            type: array
        "500":
          description: Ошибка сервера
          schema:
            type: object
      summary: Получить список всех подписок
      tags:
      - Подписки
  /read:
    post:
      consumes:
      - application/json
      description: Возвращает список подписок по заданным параметрам
      parameters:
      - description: Параметры поиска (обязательно service_name и user_id)
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/models.Subscription'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Subscription'
            type: array
        "400":
          description: Неверный формат запроса или отсутствуют обязательные поля
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Получить подписки
  /summa:
    post:
      consumes:
      - application/json
      description: Возвращает сумму подписок по заданным параметрам
      parameters:
      - description: Параметры для расчета суммы
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/models.Subscription'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RetStruct'
        "400":
          description: Неверный формат запроса или отсутствуют обязательные поля
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Расчет суммы подписок
  /update:
    put:
      consumes:
      - application/json
      description: Обновляет данные подписки в базе данных
      parameters:
      - description: 'Данные подписки для обновления (обязательные: service_name и
          user_id)'
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/models.Subscription'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RetStruct'
        "400":
          description: Неверный запрос
          schema:
            type: object
        "500":
          description: Ошибка сервера
          schema:
            type: object
      summary: Обновление подписки
swagger: "2.0"